requirements.txt
flask
numpy
pillow
opencv-python
scipy
onnxruntime
torch